renamed and refactored package to skewer
