--- pycg_bonds/__init__.py
@@ -1,1 +1,0 @@
-from .pycg_bonds import cg_bonds
